refactor scheduler
